Check if Voter is in Census before going with the next step
Changes the logic from the Mnemonic phrase as a method for generating
the Voter's wallet, to a form with some personal data:

* email
* date of birth (aka born_at): year-month-day

We use the isInCensus method from the Vocdoni SDK to check this.

Until now the PreviewVoteComponent and VoteComponent were retrieved from
the window.Decidim.setupVoteComponent object, and this worked well with
the Decidim approach of having a full page for login (aka census
checking), but this doesn't play nice with the full JavaScript approach
of the Vocdoni SDK, so I needed to change the mount logic.

This also means that we change the logic on when to use the Preview (Fake) voting:
until now this was done by just adding whichever script you needed
(setup-vote.js or setup-preview.js); now we must do it with the
`data-preview` attribute, as that allows us to send the wallet object to
the VoteComponent object in setup-vote.js.